+++ CHANGELOG.md
@@ -7,8 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- [cache] Add `disable-credential-cache` flag (breaking).
+
+### Fixed
+- [main] Prevent hang when discovery is disabled and there are no credentials or when bad credentials are given.
+
+## [0.3.1] - 2021-10-24
+
 ### Changed
 - Include build profile in the displayed version information
+- [playback] Improve dithering CPU usage by about 33%
 
 ### Fixed
 - [connect] Partly fix behavior after last track of an album/playlist
@@ -77,7 +86,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2019-11-06
 
-[unreleased]: https://github.com/librespot-org/librespot/compare/v0.3.0..HEAD
+[unreleased]: https://github.com/librespot-org/librespot/compare/v0.3.1..HEAD
+[0.3.1]: https://github.com/librespot-org/librespot/compare/v0.3.0..v0.3.1
 [0.3.0]: https://github.com/librespot-org/librespot/compare/v0.2.0..v0.3.0
 [0.2.0]: https://github.com/librespot-org/librespot/compare/v0.1.6..v0.2.0
 [0.1.6]: https://github.com/librespot-org/librespot/compare/v0.1.5..v0.1.6